Feature/improve accuracy

See merge request dune-fem/dune-vem!20

+++ dune/vem/agglomeration/agglomeration.hh
@@ -82,7 +82,7 @@ namespace Dune
       }
       void onbBasis(int order)
       {
-        if (order > maxOrder_)
+        if (order > (int)maxOrder_)
         {
           Dune::Vem::onbBasis(*this, order, boundingBoxes());
           maxOrder_ = order;
